backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
sslyze
dnspython
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic-settings
python-multipart
reportlab
numpy
matplotlib
python-dotenv
cryptography
